requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
boto3==1.28.19
botocore==1.31.19
certifi==2023.7.22
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.2.0
Django==4.2.4
django-channels==0.7.0
django-cors-headers==4.2.0
django-redis==5.3.0
django-rest-framework==0.1.0
django-storages==1.13.2
djangorestframework==3.14.0
idna==3.4
jmespath==1.0.1
msgpack==1.0.5
oauthlib==3.2.2
Pillow==10.0.0
psycopg2-binary==2.9.6
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3
redis==4.6.0
requests==2.31.0
requests-oauthlib==1.3.1
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
